requirements.txt
# Fintech AI Agent - Requirements

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Jinja2 Templates
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
python-dotenv==1.0.0

# RSS Feed Parsing
feedparser==6.0.10

# LLM APIs
openai==1.12.0

# Email Service
requests==2.31.0

# Email Sending
requests==2.31.0

# Email
# Using built-in smtplib

# HTTP Requests
requests==2.31.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
